requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==1.10.17
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
google-generativeai==0.8.3
Pillow==10.4.0

